+++ src/core/url/concat.ts
@@ -6,6 +6,7 @@
  * https://github.com/V4Fire/Core/blob/master/LICENSE
  */
 
+import { deprecate } from 'core/functools/deprecation';
 import { isAbsURL, isURLWithSlash } from 'core/url/const';
 
 /**
@@ -16,13 +17,13 @@ import { isAbsURL, isURLWithSlash } from 'core/url/const';
  * @example
  * ```js
  * // '/foo/baz/bar/bla'
- * concatUrls('foo/baz', '/bar', 'bla');
+ * concatURLs('foo/baz', '/bar', 'bla');
  *
  * // 'http://foo.bar/bla'
- * concatUrls('http://foo.bar', 'bla');
+ * concatURLs('http://foo.bar', 'bla');
  * ```
  */
-export function concatUrls(...urls: Array<Nullable<string>>): string {
+export function concatURLs(...urls: Array<Nullable<string>>): string {
 	let
 		res = '';
 
@@ -52,4 +53,19 @@ export function concatUrls(...urls: Array<Nullable<string>>): string {
 	}
 
 	return res;
+}
+
+/**
+ * @deprecated
+ * @see [[concatURLs]]
+ * @param urls
+ */
+export function concatUrls(...urls: Array<Nullable<string>>): string {
+	deprecate({
+		name: 'concatUrls',
+		type: 'function',
+		renamedTo: 'concatURLs'
+	});
+
+	return concatURLs(...urls);
 }