+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.32.1 (????-??-??)
+
+#### :boom: Breaking Change
+
+* Renamed `concatUrls` to `concatURLs` `core/url`
+
 ## v3.32.0 (2021-02-05)
 
 #### :rocket: New Feature